Add streaming via polling for Windows clients
Adds new streaming method that uses polling to request each
frame just after the last frame is received. If a requested frame
is not received within a timeout a new request is sent.

This code is being added and used for Windows clients to resolve
an issue for Windows clients where data on the USB virtual serial
link can be randomly dropped causing normal streaming to halt
and become unrecoverable due to parsing errors.