Basic Internet Protocol comms working.
This is the start of refactoring the connection classes to use
a Connection_T interface class and a factory function for creating
the appropriate type of connection based on the URI provided
to the tofcore::Sensor class.
Also introduces the IpConnection class which encapuslates the
TcpConnection and UdpServer classes for interacting with a device
over IP networks.